Stream listings of tags and branch names
Currently its a shell out command, to guard against both the message
getting too big and also the performance on the Gitaly side, a limit
will be introduced. The response is now a stream too, so we don't have
to artifically limit the number of items we send back.

Needed for https://gitlab.com/gitlab-org/gitaly/merge_requests/537#note_54781960