Create Server subpackage
Prior to this change, all proto definitions were in the root of this
project. Creating one big package, and confusion. For example, myself
and others expected each service to be a namespace, so the
RepositoyService::Exists message was a different one from
NamespaceService::Exists for example.

This change *copies* the server proto to a package directory. Having a
separte directory for each package or module/class is not a neccesity,
but it given this is the usual way of doing it in multiple languages it
seemed the cleanest way to signal a serperation of concerns.

Given the structure of the repository changed, the ruby script
generating the proto defintions has been changed too.

Changes are needed for: https://gitlab.com/gitlab-org/gitaly/issues/655