Define ListNewCommits RPC on the Ref Service
GitLab CE/EE contains a wrapper around `git-rev-list`. This hasn't been
migrated yet, and my understanding was that this wasn't required as
reuse of other RPCs would cover it.

This turned out not the be true, so a new RPC is now defined.

Links:
- https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/20087
- https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/6236

Part of: https://gitlab.com/gitlab-org/gitaly/issues/1233