Fix the Golang import path
Prior to this change, we used to import the path of the `go` directory.
This however, runs into the `go` keyword. So what we do, is alias the
package we import. This change moves it to gitalypb, inspired by the
Uber Prototool styleguide: https://github.com/uber/prototool/blob/dev/etc/style/uber/uber.proto#L38