Set ACCESSOR/MUTATOR Message options
To differentiate between mutations and operations that only access a
Gitaly server, some form of annotations are needed to achieve this. This
change includes a way to do so, on a message level.

Another option, would be to annotate the method. This might have
advantages, and instinctively it seems better to me. However, there's no
support for this right now in the golang protobuf implemenation. Also
not in newer versions that have been released.