Add operation to apply patch to a branch
GitLab rails passes the repository, user, target branch and one or
more patches.

Gitaly will create the branch if that's required on top of the branch
if it existed, create the branch from HEAD if it didn't, apply the
patches using `git am`(?).

The response will contain the resulting tip of the branch, and whether
or not a new branch was created.