Document RPC naming conventions
Gitaly proto has different naming schemes in use, and is lacking a
standard or guideline one could follow. This creates discussions which
might seem arbitrary or lack basis.

This change documents the standard as discussed in:
https://gitlab.com/gitlab-org/gitaly/issues/655. But diverging from it
regarding the Find verb, and replace it with Get. Given Get is more
commently in use right now.